Migrate column enum generator into boilerdata
