--- src/boilerdata/enums/columns.py
@@ -1,7 +1,0 @@
-from enum import auto
-
-from boilerdata.enums import NameEnum
-
-
-class Column(NameEnum):
-    T1 = auto()
